musical-time-machine/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
spotipy==2.24.0
python-dotenv~=1.0.1